Use actual war build instead of programmatic micro archive for ws tests
