+++ focused/pom.xml
@@ -220,6 +220,7 @@
                     <groupId>org.omnifaces.arquillian</groupId>
                     <artifactId>glassfish-client-ee10</artifactId>
                     <version>1.3</version>
+                    <scope>test</scope>
                 </dependency>
 
                 <!-- 
@@ -240,7 +241,6 @@
                     
                     <!-- Download and unzip GlassFish -->
                     <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
                         <artifactId>maven-dependency-plugin</artifactId>
                         <executions>
                             <execution>
@@ -268,8 +268,17 @@
                     </plugin>
                     
                     <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
                         <artifactId>maven-surefire-plugin</artifactId>
+                        <configuration>
+                            <systemPropertyVariables>
+                                <glassfish.home>${session.executionRootDirectory}/target/glassfish7</glassfish.home>
+                                <glassfish.enableDerby>true</glassfish.enableDerby>
+                            </systemPropertyVariables>
+                        </configuration>
+                    </plugin>
+                    
+                    <plugin>
+                        <artifactId>maven-failsafe-plugin</artifactId>
                         <configuration>
                             <systemPropertyVariables>
                                 <glassfish.home>${session.executionRootDirectory}/target/glassfish7</glassfish.home>
