+++ focused/websocket/onClose/src/test/java/jakartaee/examples/websocket/onclose/OnCloseEndpointIT.java
@@ -1,5 +1,5 @@
 /*
- * Permission to use, copy, modify, and/or distribute this software for any 
+ * Permission to use, copy, modify, and/or distribute this software for any
  * purpose with or without fee is hereby granted.
  *
  * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR(S) DISCLAIMS ALL WARRANTIES
@@ -10,37 +10,38 @@
  * OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN
  * CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
-package jakartaee.examples.websocket.onopen;
+package jakartaee.examples.websocket.onclose;
 
 
-import java.io.File;
+import static java.util.concurrent.TimeUnit.SECONDS;
+import static org.junit.Assert.assertTrue;
+
+import java.io.IOException;
 import java.net.URI;
 import java.net.URL;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-import jakarta.websocket.ClientEndpoint;
-import jakarta.websocket.OnMessage;
-import jakarta.websocket.Session;
-import org.jboss.arquillian.junit.Arquillian;
+
 import org.glassfish.tyrus.client.ClientManager;
-import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
+import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
-import static org.jboss.shrinkwrap.api.ShrinkWrap.create;
-import org.jboss.shrinkwrap.api.spec.WebArchive;
-import static org.junit.Assert.assertTrue;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import jakarta.websocket.ClientEndpoint;
+import jakarta.websocket.OnClose;
+import jakarta.websocket.OnMessage;
+import jakarta.websocket.Session;
+import jakartaee.examples.utils.ITBase;
+
 /**
- * A JUnit test for the @OnOpen example.
+ * A JUnit test for the @OnError example.
  *
  * @author Manfred Riem (mriem@manorrock.com)
  */
 @ClientEndpoint
 @RunWith(Arquillian.class)
-
-public class OnOpenEndpointTest {
+public class OnCloseEndpointIT extends ITBase {
 
     /**
      * Stores the base URL.
@@ -59,21 +60,6 @@ public class OnOpenEndpointTest {
     private CountDownLatch countDown = new CountDownLatch(1);
 
     /**
-     * Create the deployment web archive.
-     *
-     * @return the deployment web archive.
-     */
-    @Deployment
-    public static WebArchive createDeployment() {
-        return create(WebArchive.class).addClasses(
-                OnOpenEndpoint.class).
-                addAsWebResource(new File("src/main/webapp/index.xhtml")).
-                addAsWebInfResource(new File("src/main/webapp/WEB-INF/web.xml")).
-                addAsWebInfResource(new File("src/main/webapp/WEB-INF/beans.xml"))
-                ;
-    }
-
-    /**
      * Get the buffer.
      *
      * @return the buffer.
@@ -87,10 +73,22 @@ public class OnOpenEndpointTest {
      *
      * @param session the session.
      * @param message the message.
+     * @throws Exception when a serious error occurs.
      */
     @OnMessage
-    public void onMessage(Session session, String message) {
-        buffer.append(message);
+    public void onMessage(Session session, String message) throws Exception {
+        session.getBasicRemote().sendText(message);
+    }
+
+    /**
+     * Handle the onClose.
+     *
+     * @param session the session.
+     * @throws IOException when an I/O error occurs.
+     */
+    @OnClose
+    public void onClose(Session session) throws IOException {
+        buffer.append("Closing connection");
         countDown.countDown();
     }
 
@@ -103,12 +101,20 @@ public class OnOpenEndpointTest {
     @RunAsClient
     public void testClientEndpoint() throws Exception {
         countDown = new CountDownLatch(1);
+
         ClientManager client = ClientManager.createClient();
         StringBuilder wsUrl = new StringBuilder();
-        wsUrl.append("ws://").append(baseUrl.getHost()).append(":").
-                append(baseUrl.getPort()).append(baseUrl.getPath()).append("echo");
+        wsUrl.append("ws://")
+             .append(baseUrl.getHost())
+             .append(":")
+             .append(baseUrl.getPort())
+             .append(baseUrl.getPath())
+             .append("echo");
+
         client.connectToServer(this, new URI(wsUrl.toString()));
-        countDown.await(100, TimeUnit.SECONDS);
-        assertTrue(buffer.toString().contains("Session started at "));
+
+        countDown.await(10, SECONDS);
+        System.out.println(buffer.toString());
+        assertTrue(buffer.toString().contains("Closing connection"));
     }
 }
