+++ focused/websocket/onError/src/test/java/jakartaee/examples/websocket/onerror/OnErrorEndpointIT.java
@@ -1,5 +1,5 @@
 /*
- * Permission to use, copy, modify, and/or distribute this software for any 
+ * Permission to use, copy, modify, and/or distribute this software for any
  * purpose with or without fee is hereby granted.
  *
  * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR(S) DISCLAIMS ALL WARRANTIES
@@ -10,39 +10,37 @@
  * OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN
  * CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
-package jakartaee.examples.websocket.annotatedclientendpoint;
+package jakartaee.examples.websocket.onerror;
 
-import java.io.File;
-import java.io.IOException;
+
+import static org.junit.Assert.assertTrue;
+
 import java.net.URI;
 import java.net.URL;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
-import jakarta.websocket.ClientEndpoint;
-import jakarta.websocket.OnMessage;
-import jakarta.websocket.OnOpen;
-import jakarta.websocket.Session;
-import org.glassfish.tyrus.client.ClientManager;
-import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.container.test.api.RunAsClient;
 
+import org.glassfish.tyrus.client.ClientManager;
+import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
-import static org.jboss.shrinkwrap.api.ShrinkWrap.create;
-import org.jboss.shrinkwrap.api.spec.WebArchive;
-import static org.junit.Assert.assertEquals;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import jakarta.websocket.ClientEndpoint;
+import jakarta.websocket.OnError;
+import jakarta.websocket.OnMessage;
+import jakarta.websocket.Session;
+import jakartaee.examples.utils.ITBase;
+
 /**
- * An annotated ClientEndpoint for the annotated ClientEndpoint example.
+ * A JUnit test for the @OnError example.
  *
  * @author Manfred Riem (mriem@manorrock.com)
  */
 @ClientEndpoint
 @RunWith(Arquillian.class)
-
-public class AnnotatedClientEndpointTest {
+public class OnErrorEndpointIT extends ITBase {
 
     /**
      * Stores the base URL.
@@ -61,21 +59,6 @@ public class AnnotatedClientEndpointTest {
     private CountDownLatch countDown = new CountDownLatch(1);
 
     /**
-     * Create the deployment web archive.
-     *
-     * @return the deployment web archive.
-     */
-    @Deployment
-    public static WebArchive createDeployment() {
-        return create(WebArchive.class).addClasses(
-                AnnotatedClientEndpointServerEndpoint.class).
-                addAsWebResource(new File("src/main/webapp/index.xhtml")).
-                addAsWebInfResource(new File("src/main/webapp/WEB-INF/web.xml")).
-                addAsWebInfResource(new File("src/main/webapp/WEB-INF/beans.xml"))
-                ;
-    }
-
-    /**
      * Get the buffer.
      *
      * @return the buffer.
@@ -85,17 +68,14 @@ public class AnnotatedClientEndpointTest {
     }
 
     /**
-     * Handle the on open event.
+     * Handle the OnError.
      *
-     * @param session the session.
+     * @param cause the cause.
      */
-    @OnOpen
-    public void onOpen(Session session) {
-        try {
-            session.getBasicRemote().sendText("ECHO");
-        } catch (IOException ioe) {
-            buffer.append(ioe.getMessage());
-        }
+    @OnError
+    public void onError(Throwable cause) {
+        buffer.append(cause.getMessage());
+        countDown.countDown();
     }
 
     /**
@@ -103,11 +83,12 @@ public class AnnotatedClientEndpointTest {
      *
      * @param session the session.
      * @param message the message.
+     * @throws Exception when a serious error occurs.
      */
     @OnMessage
-    public void onMessage(Session session, String message) {
-        buffer.append(message);
-        countDown.countDown();
+    public void onMessage(Session session, String message) throws Exception {
+        session.close();
+        session.getBasicRemote().sendText(message);
     }
 
     /**
@@ -125,6 +106,7 @@ public class AnnotatedClientEndpointTest {
                 append(baseUrl.getPort()).append(baseUrl.getPath()).append("echo");
         client.connectToServer(this, new URI(wsUrl.toString()));
         countDown.await(100, TimeUnit.SECONDS);
-        assertEquals("ECHO", buffer.toString());
+        System.out.println(buffer.toString());
+        assertTrue(buffer.toString().contains("The connection has been closed."));
     }
 }
