+++ focused/websocket/onOpen/src/test/java/jakartaee/examples/websocket/onopen/OnOpenEndpointIT.java
@@ -1,5 +1,5 @@
 /*
- * Permission to use, copy, modify, and/or distribute this software for any 
+ * Permission to use, copy, modify, and/or distribute this software for any
  * purpose with or without fee is hereby granted.
  *
  * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR(S) DISCLAIMS ALL WARRANTIES
@@ -10,38 +10,36 @@
  * OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN
  * CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
-package jakartaee.examples.websocket.onerror;
+package jakartaee.examples.websocket.onopen;
 
 
-import java.io.File;
+import static java.util.concurrent.TimeUnit.SECONDS;
+import static org.junit.Assert.assertTrue;
+
 import java.net.URI;
 import java.net.URL;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-import jakarta.websocket.ClientEndpoint;
-import jakarta.websocket.OnError;
-import jakarta.websocket.OnMessage;
-import jakarta.websocket.Session;
-import org.jboss.arquillian.junit.Arquillian;
+
 import org.glassfish.tyrus.client.ClientManager;
-import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.RunAsClient;
+import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
-import static org.jboss.shrinkwrap.api.ShrinkWrap.create;
-import org.jboss.shrinkwrap.api.spec.WebArchive;
-import static org.junit.Assert.assertTrue;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import jakarta.websocket.ClientEndpoint;
+import jakarta.websocket.OnMessage;
+import jakarta.websocket.Session;
+import jakartaee.examples.utils.ITBase;
+
 /**
- * A JUnit test for the @OnError example.
+ * A JUnit test for the @OnOpen example.
  *
  * @author Manfred Riem (mriem@manorrock.com)
  */
 @ClientEndpoint
 @RunWith(Arquillian.class)
-
-public class OnErrorEndpointTest {
+public class OnOpenEndpointIT extends ITBase {
 
     /**
      * Stores the base URL.
@@ -60,21 +58,6 @@ public class OnErrorEndpointTest {
     private CountDownLatch countDown = new CountDownLatch(1);
 
     /**
-     * Create the deployment web archive.
-     *
-     * @return the deployment web archive.
-     */
-    @Deployment
-    public static WebArchive createDeployment() {
-        return create(WebArchive.class).addClasses(
-                OnErrorEndpoint.class).
-                addAsWebResource(new File("src/main/webapp/index.xhtml")).
-                addAsWebInfResource(new File("src/main/webapp/WEB-INF/web.xml")).
-                addAsWebInfResource(new File("src/main/webapp/WEB-INF/beans.xml"))
-                ;
-    }
-
-    /**
      * Get the buffer.
      *
      * @return the buffer.
@@ -82,29 +65,17 @@ public class OnErrorEndpointTest {
     public String getBuffer() {
         return buffer.toString();
     }
-    
-    /**
-     * Handle the OnError.
-     * 
-     * @param cause the cause.
-     */
-    @OnError
-    public void onError(Throwable cause) {
-        buffer.append(cause.getMessage());
-        countDown.countDown();
-    }
-    
+
     /**
      * Handle the text message.
      *
      * @param session the session.
      * @param message the message.
-     * @throws Exception when a serious error occurs.
      */
     @OnMessage
-    public void onMessage(Session session, String message) throws Exception {
-        session.close();
-        session.getBasicRemote().sendText(message);
+    public void onMessage(Session session, String message) {
+        buffer.append(message);
+        countDown.countDown();
     }
 
     /**
@@ -116,13 +87,20 @@ public class OnErrorEndpointTest {
     @RunAsClient
     public void testClientEndpoint() throws Exception {
         countDown = new CountDownLatch(1);
+
         ClientManager client = ClientManager.createClient();
         StringBuilder wsUrl = new StringBuilder();
-        wsUrl.append("ws://").append(baseUrl.getHost()).append(":").
-                append(baseUrl.getPort()).append(baseUrl.getPath()).append("echo");
+        wsUrl.append("ws://")
+             .append(baseUrl
+             .getHost())
+             .append(":")
+             .append(baseUrl.getPort())
+             .append(baseUrl.getPath())
+             .append("echo");
+
         client.connectToServer(this, new URI(wsUrl.toString()));
-        countDown.await(100, TimeUnit.SECONDS);
-        System.out.println(buffer.toString());
-        assertTrue(buffer.toString().contains("The connection has been closed."));
+
+        countDown.await(10, SECONDS);
+        assertTrue(buffer.toString().contains("Session started at "));
     }
 }
